First cut of using setuptools for installation
